Move path utilities to separate package
