+++ change/@ws-tools-paths-d8fc65a9-0286-4b6a-9e09-686e024b882b.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "Create package",
+  "packageName": "@ws-tools/paths",
+  "email": "elcraig@microsoft.com",
+  "dependentChangeType": "patch"
+}
