+++ change/workspace-tools-14b42056-f741-41ac-b7ad-c5079d191ab9.json
@@ -1,0 +1,7 @@
+{
+  "type": "minor",
+  "comment": "Move path utilities to separate package (re-exported by this one)",
+  "packageName": "workspace-tools",
+  "email": "elcraig@microsoft.com",
+  "dependentChangeType": "patch"
+}
