+++ packages/paths/README.md
@@ -1,0 +1,5 @@
+# @ws-tools/paths
+
+Helper methods related to paths in a repo or monorepo. The methods in this package are re-exported from [`workspace-tools`](https://www.npmjs.com/package/workspace-tools) for convenience.
+
+This package should contain only the most basic helpers with no or minimal dependencies.
