+++ packages/paths/etc/paths.api.md
@@ -1,0 +1,18 @@
+## API Report File for "@ws-tools/paths"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+
+// @public
+export function findGitRoot(cwd: string): string;
+
+// @public
+export function findPackageRoot(cwd: string): string | undefined;
+
+// @public
+export function searchUp(filePath: string | string[], cwd: string): string | undefined;
+
+// (No @packageDocumentation comment for this package)
+
+```
