+++ packages/paths/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "workspace-tools",
-  "version": "0.34.3",
+  "name": "@ws-tools/paths",
+  "version": "0.1.0",
   "license": "MIT",
   "repository": {
     "type": "git",
@@ -10,7 +10,7 @@
   "types": "lib/index.d.ts",
   "files": [
     "lib/!(__*)",
-    "lib/!(__*)/**/*"
+    "lib/!(__*)/**"
   ],
   "scripts": {
     "api": "workspace-tools-scripts api",
@@ -18,17 +18,7 @@
     "start": "tsc -w --preserveWatchOutput",
     "test": "jest"
   },
-  "dependencies": {
-    "@yarnpkg/lockfile": "^1.1.0",
-    "git-url-parse": "^13.0.0",
-    "globby": "^11.0.0",
-    "jju": "^1.4.0",
-    "js-yaml": "^4.1.0",
-    "micromatch": "^4.0.0",
-    "@ws-tools/paths": "^0.1.0"
-  },
   "devDependencies": {
-    "lodash": "4.17.21",
     "workspace-tools-scripts": "*"
   }
 }
